Rename fragments and api functions
