+++ app/src/main/java/com/innowisegroup/datetimepicker/DatePickerFragment.kt
@@ -7,11 +7,11 @@ import android.view.ViewGroup
 import android.widget.TextView
 import androidx.core.content.res.ResourcesCompat
 import androidx.fragment.app.Fragment
-import com.innowisegroup.datetimepicker.PickerDialog.Companion.UPDATE_DATE_TAB_TITLE_KEY
-import com.innowisegroup.datetimepicker.PickerDialog.Companion.UPDATE_DATE_TAB_TITLE_REQUEST_KEY
+import com.innowisegroup.datetimepicker.ReelPicker.Companion.UPDATE_DATE_TAB_TITLE_KEY
+import com.innowisegroup.datetimepicker.ReelPicker.Companion.UPDATE_DATE_TAB_TITLE_REQUEST_KEY
 import java.util.*
 
-class FragmentDatePicker : Fragment() {
+class DatePickerFragment : Fragment() {
 
     var day: CustomNumberPicker? = null
     var month: CustomNumberPicker? = null
